Move test_marker_base_operator to helpers
